fastapi_pneumolia_detect/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.103.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.5
h11==0.14.0
h5py==3.11.0
idna==3.7
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
pip==24.2
protobuf==4.25.4
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
python-multipart==0.0.9
requests==2.32.3
rich==13.7.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
typing_extensions==4.11.0
urllib3==2.2.2
uvicorn==0.20.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
